+++ conf/script/src/build_system/cmd/hierarchy/find_root_dir/__init__.py
@@ -1,1 +1,1 @@
-from build_system.cmd.hierarchy.find_root_dir.find_root_dir import find_root_dir
+from build_system.cmd.hierarchy.find_root_dir.impl import find_root_dir
