+++ conf/script/src/build_system/cmd/hierarchy/find_root_dir/cli.py
@@ -3,7 +3,7 @@ import argparse
 import utils.error.cls_def
 import utils.error.managed
 from build_system import cmd
-from build_system.cmd.hierarchy.find_root_dir.find_root_dir import BUILD_SYSTEM_CONF_FILENAME
+from build_system.cmd.hierarchy.find_root_dir.impl import BUILD_SYSTEM_CONF_FILENAME
 
 
 def find_root_dir():
